refactor: Move some state and computed value to `GraphStore`
